Switch to FP16 as FP32 is not available anymore...

+++ python/hugging_face_diffusers/sdxl-1.0-base_cpu.py
@@ -5,7 +5,7 @@
 from diffusers import DiffusionPipeline
 import torch
 
-pipe = DiffusionPipeline.from_pretrained("stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float32, use_safetensors=True, variant="fp32")
+pipe = DiffusionPipeline.from_pretrained("stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float16, use_safetensors=True, variant="fp16")
 pipe.to("cpu")
 
 # if using torch < 2.0
